requirements.txt
autopep8==1.5.4
mysql-connector-python==8.0.22
pkg-resources==0.0.0
prettytable==2.0.0
protobuf==3.14.0
pycodestyle==2.6.0
six==1.15.0
toml==0.10.2
wcwidth==0.2.5
